Generation functions moved to workflows
